hide refresh visualization button in visitor mode
