+++ src/app/models/state.model.ts
@@ -27,7 +27,6 @@ export interface AppState {
   expertMode: boolean;
   communityWorkspace: CommunityWorkspace;
   currentWorkSpaceOwner: string;
-  restricted: boolean;
   requirements: Requirement[];
 }
 
@@ -48,7 +47,6 @@ export const INITIAL_APP_STATE: AppState = {
   expertMode: false,
   communityWorkspace: undefined,
   currentWorkSpaceOwner: undefined,
-  restricted: false,
   requirements: undefined,
 };
 /**
