+++ src/app/models/workspace.model.ts
@@ -2,7 +2,7 @@ import { MeasureCatalog } from './measure.catalog';
 import { Requirement } from './reqbaz.model';
 import { ServiceInformation } from './service.model';
 import { SuccessModel } from './success.model';
-import { Visitor } from './user.model';
+
 import { VisualizationCollection } from './visualization.model';
 
 export interface ApplicationWorkspace {
@@ -23,4 +23,13 @@ export interface UserWorkspace {
 export interface CommunityWorkspace {
   // user ID is key
   [key: string]: UserWorkspace;
+}
+export class Visitor {
+  constructor(public username: string, public role: UserRole) {}
+}
+
+export enum UserRole {
+  SPECTATOR = 'spectator',
+  EDITOR = 'editor',
+  LURKER = 'lurker',
 }
