+++ src/app/visualizations/kpi-visualization/kpi-visualization.component.html
@@ -6,18 +6,19 @@
       <button mat-icon-button [matMenuTriggerFor]="menu">
         <mat-icon> info_outline </mat-icon>
       </button>
-
-      <button
-        *ngIf="query$ | async as query"
-        mat-icon-button
-        matTooltipShowDelay="0"
-        matTooltip="{{
-          'success-modeling.visualization.refresh-data' | translate
-        }}"
-        (click)="onRefreshClicked(query)"
+      <ng-container *ngIf="!(restricted$ | async)">
+        <button
+          *ngIf="query$ | async as query"
+          mat-icon-button
+          matTooltipShowDelay="0"
+          matTooltip="{{
+            'success-modeling.visualization.refresh-data' | translate
+          }}"
+          (click)="onRefreshClicked(query)"
+        >
+          <mat-icon color="primary">sync</mat-icon>
+        </button></ng-container
       >
-        <mat-icon color="primary">sync</mat-icon>
-      </button>
       <mat-menu #menu="matMenu" yPosition="above">
         <div id="chart-info">
           <app-visualization-info
